refactor(api): separate event service to smaller services
- Removed deprecated `is_seat_pickable` property from seat sections in seat.service.ts.
- Added new seatmap.service.ts to handle seatmap-related operations, including bulk insertion of sections and seats for shows.
- Created show.service.ts to manage show-related functionalities, including adding, updating, and deleting shows.
- Updated event service references in API routes to use the new seatmap and show services.
- Improved ticket code generation with cryptographically secure randomness in ticket-code.ts.
- Cleaned up error handling in checkout routes for orders and events.
- Adjusted schemas to remove deprecated fields and ensure consistency across the application.
