+++ src/lib/server/errors.ts
@@ -63,16 +63,28 @@ export const Errors = {
   ORDER_ALREADY_PROCESSED: new AppError('ORDER_ALREADY_PROCESSED', 400, 'Đơn hàng đã được xử lý'),
   LOCK_EXPIRED: new AppError('LOCK_EXPIRED', 410, 'Đơn hàng đã hết hạn, vui lòng đặt lại vé'),
   ORDER_EMPTY: new AppError('ORDER_EMPTY', 400, 'Đơn hàng không có vé nào'),
-  ACTIVE_ORDER_EXISTS: new AppError('ACTIVE_ORDER_EXISTS', 409, 'Bạn đang có đơn hàng đang chờ xử lý'),
+  ACTIVE_ORDER_EXISTS: new AppError(
+    'ACTIVE_ORDER_EXISTS',
+    409,
+    'Bạn đang có đơn hàng đang chờ xử lý',
+  ),
 
   // Seats
   SEAT_NOT_AVAILABLE: new AppError('SEAT_NOT_AVAILABLE', 409, 'Một hoặc nhiều ghế đã được đặt'),
   SALES_NOT_STARTED: new AppError('SALES_NOT_STARTED', 400, 'Chưa đến thời gian mở bán'),
   SALES_ENDED: new AppError('SALES_ENDED', 400, 'Đã hết thời gian bán vé'),
-  MAX_TICKETS_EXCEEDED: new AppError('MAX_TICKETS_EXCEEDED', 400, 'Vượt quá số lượng vé tối đa cho phép'),
+  MAX_TICKETS_EXCEEDED: new AppError(
+    'MAX_TICKETS_EXCEEDED',
+    400,
+    'Vượt quá số lượng vé tối đa cho phép',
+  ),
   HOLD_FAILED: new AppError('HOLD_FAILED', 500, 'Không thể giữ ghế, vui lòng thử lại'),
   SEATS_LIST_EMPTY: new AppError('SEATS_LIST_EMPTY', 400, 'Danh sách ghế không được để trống'),
-  MAX_SEATS_EXCEEDED: new AppError('MAX_SEATS_EXCEEDED', 400, 'Vượt quá số lượng ghế tối đa cho phép'),
+  MAX_SEATS_EXCEEDED: new AppError(
+    'MAX_SEATS_EXCEEDED',
+    400,
+    'Vượt quá số lượng ghế tối đa cho phép',
+  ),
   INVALID_SECTION_TYPE: new AppError('INVALID_SECTION_TYPE', 400, 'Ghế không thuộc đúng khu'),
   DUPLICATE_SEAT: new AppError('DUPLICATE_SEAT', 400, 'Ghế bị trùng'),
   INVALID_QUANTITY: new AppError('INVALID_QUANTITY', 400, 'Số lượng ghế đứng không hợp lệ'),
@@ -83,7 +95,11 @@ export const Errors = {
   DUPLICATE_SHOW: new AppError('DUPLICATE_SHOW', 400, 'Suất diễn bị trùng'),
 
   // Idempotency
-  IDEMPOTENCY_CONFLICT: new AppError('IDEMPOTENCY_CONFLICT', 409, 'Yêu cầu idempotency đang xung đột'),
+  IDEMPOTENCY_CONFLICT: new AppError(
+    'IDEMPOTENCY_CONFLICT',
+    409,
+    'Yêu cầu idempotency đang xung đột',
+  ),
 
   // General
   NOT_FOUND: new AppError('NOT_FOUND', 404, 'Không tìm thấy'),
@@ -95,7 +111,12 @@ export const Errors = {
     new AppError('VALIDATION_ERROR', 400, 'Dữ liệu không hợp lệ', details),
 
   CART_CONFLICT: (details?: Record<string, string>) =>
-    new AppError( 'CART_CONFLICT', 409, 'Một số vé trong giỏ hàng đã bị người khác mua hoặc không đủ số lượng', details),
+    new AppError(
+      'CART_CONFLICT',
+      409,
+      'Một số vé trong giỏ hàng đã bị người khác mua hoặc không đủ số lượng',
+      details,
+    ),
 } as const;
 
 /**
