+++ src/routes/api/events/[id]/shows/[showId]/+server.ts
@@ -1,6 +1,7 @@
 import { requireAdmin } from '$lib/server/auth/guards';
 import { apiHandler } from '$lib/server/handler';
-import { eventService } from '$lib/server/services/event.service';
+import { seatmapService } from '$lib/server/services/seatmap.service';
+import { showService } from '$lib/server/services/show.service';
 import { json } from '@sveltejs/kit';
 
 /** PUT — Replace all sections (and their seats) for a specific show */
@@ -8,7 +9,7 @@ export const PUT = apiHandler(async ({ request, params, locals }) => {
   const admin = requireAdmin(locals);
   const body = await request.json();
   const showId = Number((params as Record<string, string>).showId);
-  const data = await eventService.updateShowSections(admin.id, showId, body);
+  const data = await seatmapService.updateShowSections(admin.id, showId, body);
   return json({ data }, { status: 200 });
 });
 
@@ -17,7 +18,7 @@ export const PATCH = apiHandler(async ({ request, params, locals }) => {
   const admin = requireAdmin(locals);
   const body = await request.json();
   const showId = (params as Record<string, string>).showId;
-  const data = await eventService.updateShow(admin.id, showId, body);
+  const data = await showService.updateShow(admin.id, showId, body);
   return json({ data });
 });
 
@@ -25,6 +26,6 @@ export const PATCH = apiHandler(async ({ request, params, locals }) => {
 export const DELETE = apiHandler(async ({ params, locals }) => {
   const admin = requireAdmin(locals);
   const showId = (params as Record<string, string>).showId;
-  const data = await eventService.deleteShow(admin.id, showId);
+  const data = await showService.deleteShow(admin.id, showId);
   return json({ data });
 });
