+++ src/routes/api/events/create/seatmap/+server.ts
@@ -1,6 +1,6 @@
 import { requireAdmin } from '$lib/server/auth/guards';
 import { apiHandler } from '$lib/server/handler';
-import { eventService } from '$lib/server/services/event.service';
+import { seatmapService } from '$lib/server/services/seatmap.service';
 import { json } from '@sveltejs/kit';
 
 /**
@@ -11,6 +11,6 @@ import { json } from '@sveltejs/kit';
 export const POST = apiHandler(async ({ request, locals }) => {
   const admin = requireAdmin(locals);
   const body = await request.json();
-  const data = await eventService.addSeatmap(admin.id, body);
+  const data = await seatmapService.addSeatmap(admin.id, body);
   return json({ data }, { status: 201 });
 });
